python-backend/functions/infer/src/requirements.txt
boto3==1.36.0
Pillow==11.1.0
# transformers==4.48.0
